allow a connection to be added thru sessionConnections, fix integration, next work on canconfigure bool

+++ products/jbrowse-web/src/sessionModelFactory.ts
@@ -179,14 +179,10 @@ export default function sessionModelFactory(pluginManager: PluginManager) {
       ) {
         const { type } = configuration
         if (!type) throw new Error('track configuration has no `type` listed')
-        const name = self.adminMode
-          ? readConfObject(configuration, 'name')
-          : configuration.name
+        const name = readConfObject(configuration, 'name')
         const connectionType = pluginManager.getConnectionType(type)
         if (!connectionType) throw new Error(`unknown connection type ${type}`)
-        const assemblyName = self.adminMode
-          ? readConfObject(configuration, 'assemblyName')
-          : configuration.assemblyName
+        const assemblyName = readConfObject(configuration, 'assemblyName')
         const connectionData = {
           ...initialSnapshot,
           name,
@@ -199,6 +195,7 @@ export default function sessionModelFactory(pluginManager: PluginManager) {
         if (!assemblyConnections)
           throw new Error(`assembly ${assemblyName} not found`)
         const length = assemblyConnections.push(connectionData)
+        console.log('here', JSON.stringify(assemblyConnections))
         return assemblyConnections[length - 1]
       },
 
@@ -247,12 +244,8 @@ export default function sessionModelFactory(pluginManager: PluginManager) {
       prepareToBreakConnection(configuration: AnyConfigurationModel) {
         const callbacksToDereferenceTrack: Function[] = []
         const dereferenceTypeCount: Record<string, number> = {}
-        const name = self.adminMode
-          ? readConfObject(configuration, 'name')
-          : configuration.name
-        const assemblyName = self.adminMode
-          ? readConfObject(configuration, 'assemblyName')
-          : configuration.assemblyName
+        const name = readConfObject(configuration, 'name')
+        const assemblyName = readConfObject(configuration, 'assemblyName')
         const assemblyConnections =
           self.connectionInstances.get(assemblyName) || []
         const connection = assemblyConnections.find(c => c.name === name)
@@ -276,12 +269,8 @@ export default function sessionModelFactory(pluginManager: PluginManager) {
       },
 
       breakConnection(configuration: AnyConfigurationModel) {
-        const name = self.adminMode
-          ? readConfObject(configuration, 'name')
-          : configuration.name
-        const assemblyName = self.adminMode
-          ? readConfObject(configuration, 'assemblyName')
-          : configuration.assemblyName
+        const name = readConfObject(configuration, 'name')
+        const assemblyName = readConfObject(configuration, 'assemblyName')
         const connectionInstances = self.connectionInstances.get(assemblyName)
         if (!connectionInstances)
           throw new Error(`connections for ${assemblyName} not found`)
@@ -290,7 +279,17 @@ export default function sessionModelFactory(pluginManager: PluginManager) {
       },
 
       deleteConnection(configuration: AnyConfigurationModel) {
-        return getParent(self).jbrowse.deleteConnectionConf(configuration)
+        if (self.adminMode) {
+          return getParent(self).jbrowse.deleteConnectionConf(configuration)
+        }
+        const { connectionId } = configuration
+        const idx = self.sessionConnections.findIndex(
+          c => c.connectionId === connectionId,
+        )
+        if (idx === -1) {
+          return undefined
+        }
+        return self.sessionConnections.splice(idx, 1)
       },
 
       updateDrawerWidth(drawerWidth: number) {
@@ -381,13 +380,13 @@ export default function sessionModelFactory(pluginManager: PluginManager) {
         if (!type) {
           throw new Error(`unknown connection type ${type}`)
         }
-
-        const connection = self.sessionConnections.find(
+        const connection = self.sessionTracks.find(
           (c: any) => c.connectionId === connectionId,
         )
-        if (connection) return connection
-
-        const length = self.sessionConnections.push(connection)
+        if (connection) {
+          return connection
+        }
+        const length = self.sessionConnections.push(connectionConf)
         return self.sessionConnections[length - 1]
       },
